First REST service /dealerLists implemented without Angular

+++ PioneerCylinderTracker/src/main/java/com/pio/PioneerCylinderTracker/model/DealerDetailsBean.java
@@ -46,7 +46,7 @@ public class DealerDetailsBean {
 	private Timestamp dateOfRegistration;
 	
 	@NonNull
-	private long totalCylindersTaken;
+	private Integer totalCylindersTaken;
 	
 	
 }
